% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoders.R
\name{find_address_candidates}
\alias{find_address_candidates}
\title{Find address candidates for City of Dallas addresses}
\usage{
find_address_candidates(
  street,
  city = NULL,
  zip = NULL,
  max_locs = NULL,
  server = c("DallasStreetsLocator", "ParcelLocator", "AccountPointsLocator",
    "AccountpointsStreetLocator")
)
}
\arguments{
\item{street}{street address to match against. Must be a single string (i.e.
a character vector of length 1)}

\item{city}{optional, city to match against. Must be a single string (i.e.
a character vector of length 1)}

\item{zip}{optional, postal code to match against. Must be a single string
(i.e. a character vector of length 1)}

\item{max_locs}{optional, maximum number of address candidates to return.
Must be a single scalar value.}

\item{server}{GeocodeServer to use. Defaults to Dallas StreetsLocator. Other
options include AccountPointsLocator, AccountpointsStreetLocator, and
ParcelLocator. See the corresponding pages under the
\href{https://gis.dallascityhall.com/wwwgis/rest/services/ToolServices}{
ToolServices} directory for additional details.}
}
\value{
a \code{data.frame} of all address candidates and respective lat/long
coordinates and address matching scores
}
\description{
Identify matching City of Dallas addresses for a given address using the
\href{https://gis.dallascityhall.com/wwwgis/rest/services/ToolServices/
DallasStreetsLocator/GeocodeServer/reverseGeocode}{City of Dallas's public
address candidate service}. Note that this operation can only process one
address at a time. See ArcGIS's
\href{https://developers.arcgis.com/rest/geocode/api-reference/
geocoding-find-address-candidates.htm}{findAddressCandidates} documentation
page for additional details
}
\examples{
find_address_candidates("1500 Marilla St", max_locs = 1)

}
