% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoders.R
\name{geocode_addresses}
\alias{geocode_addresses}
\title{Geocode addresses using the City of Dallas's public geocoding service}
\usage{
geocode_addresses(street, city = NULL, zip = NULL, id = NULL)
}
\arguments{
\item{street}{character vector of street addresses.}

\item{city}{optional, character vector of city names.}

\item{zip}{optional, character vector of postal codes.}

\item{id}{optional, numeric vector uniquely identifying each address,
defaults to \code{seq(street)} if not provided. If providing this
argument, note that each integer in the vector must be unique.}
}
\value{
A \code{data.frame} of geocoded address results.
}
\description{
Calculate lat/long coordinates for a given set of addresses using the City of
Dallas's public geocoding service. Note that this service can geocode a
\strong{maximum 1000 addresses} per request. Calling this function on more
than 1000 addresses will return an error. See the
\href{https://gis.dallascityhall.com/wwwgis/rest/services/ToolServices/
DallasStreetsLocator/GeocodeServer}{ArcGIS REST API endpoint} for additional
details.
}
\examples{
addresses <- data.frame(
  street = c('8525 Garland Rd', '1500 Marilla St', '3809 Grand Avenue'),
  city = c('Dallas', 'Dallas', 'Dallas'),
  zip = c('75218', '75201', '75210')
)

geocode_addresses(addresses$street, addresses$city, addresses$zip)
geocode_addresses(addresses$street)
}
