% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoders.R
\name{geocode_addresses}
\alias{geocode_addresses}
\title{Convert City of Dallas addresses to lat/long coordinates}
\usage{
geocode_addresses(
  street,
  city = NULL,
  zip = NULL,
  id = seq(street),
  server = c("DallasStreetsLocator", "ParcelLocator", "AccountPointsLocator",
    "AccountpointsStreetLocator"),
  out_sr = 4326,
  output = c("latlong", "all")
)
}
\arguments{
\item{street}{character vector of street addresses to geocode.}

\item{city}{optional, character vector of city names to geocode.}

\item{zip}{optional, character vector of postal codes to geocode.}

\item{id}{optional, numeric vector uniquely identifying each address, If
providing this argument (e.g., to join onto an existing dataframe), note
that each integer must be unique.}

\item{server}{GeocodeServer to use. Defaults to DallasStreetsLocator.
Refer to the \href{https://gis.dallascityhall.com/wwwgis/rest/services/
ToolServices}{ToolServices} API directory for details on each server.}

\item{out_sr}{Spatial Reference ID (WKID) for the desired lat/long
coordinate system. Defaults to 4326, which corresponds to
\href{https://wiki.gis.com/wiki/index.php/WGS84}{WGS84}. If the coordinates
will be used with the City of Dallas's MapServers, you may wish to set this
to 2276, the default WKID for many Dallas services. Refer to the
\href{https://developers.arcgis.com/rest/services-reference/enterprise/
projected-coordinate-systems.htm}{Projected} and \href{https://developers.
arcgis.com/rest/services-reference/enterprise/geographic-coordinate-systems
.htm}{Geographic} coordinate system documentation pages for a full list of
valid WKIDs.}

\item{output}{Function output. If \code{latlong} (default), function will
return a dataframe of lat/long coordinates, addresses, and a few other
fields. If \code{all}, will return a list corresponding to the full JSON
response from the server.}
}
\value{
A dataframe or list of geocoded address results.
}
\description{
Calculate lat/long coordinates for a given set of addresses using the
City of Dallas's public \href{https://gis.dallascityhall.com/wwwgis/rest/
services/ToolServices/DallasStreetsLocator/GeocodeServer/geocodeAddresses}{
geocodeAddresses service}. Note that this service can geocode a
\strong{maximum 1000 addresses} per function call. See ArcGIS's
\href{https://developers.arcgis.com/rest/geocode/api-reference/
geocoding-geocode-addresses.htm}{documentation page} for additional details.
}
\examples{
addresses <- data.frame(
  street = c('8525 Garland Rd', '1500 Marilla St', '3809 Grand Avenue'),
  city = c('Dallas', 'Dallas', 'Dallas'),
  zip = c('75218', '75201', '75210')
)

geocode_addresses(addresses$street, addresses$city, addresses$zip)
geocode_addresses(addresses$street)

}
