% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoders.R
\name{reverse_geocode}
\alias{reverse_geocode}
\title{Reverse Geocode addresses using the City of Dallas's public geocoding service}
\usage{
reverse_geocode(latitude, longitude, intersection = F, distance = 0)
}
\arguments{
\item{latitude}{scalar, latitude coordinate to reverse geocode,}

\item{longitude}{scalar, longitude coordinate to reverse geocode}

\item{intersection}{logical boolean value specifying whether the
geocode service should return the nearest street intersection or the
nearest address to the given point. Default is FALSE.}

\item{distance}{scalar, maximum distance in meters from coordinates
within which a matching address should be searched. Default value is 0.}
}
\value{
A single row \code{data.frame} of the reverse-geocoded address with
columns /code{street}, /code{city}, and /code{zip}
}
\description{
Determine addresses from lat/long coordinates using the City of Dallas's
public geocoding service. See the
\href{https://gis.dallascityhall.com/wwwgis/rest/services/ToolServices/
DallasStreetsLocator/GeocodeServer}{ArcGIS REST API endpoint} for additional
details.
}
