% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoders.R
\name{reverse_geocode}
\alias{reverse_geocode}
\title{Convert lat/long coordinates to City of Dallas addresses}
\usage{
reverse_geocode(latitude, longitude, intersection = F, distance = 0)
}
\arguments{
\item{latitude}{scalar, latitude coordinate to reverse geocode,}

\item{longitude}{scalar, longitude coordinate to reverse geocode}

\item{intersection}{logical boolean value specifying whether the
geocode service should return the nearest street intersection or the
nearest address to the given point. Default is FALSE.}

\item{distance}{scalar, maximum distance in meters from coordinates
within which a matching address should be searched. Default value is 0.}
}
\value{
A single row \code{data.frame} of the reverse-geocoded address with
columns \code{street}, \code{city}, and \code{zip}
}
\description{
Convert lat/long coordinates to City of Dallas addresses using the
\href{https://gis.dallascityhall.com/wwwgis/rest/services/ToolServices/
DallasStreetsLocator/GeocodeServer/reverseGeocode}{City of Dallas's public
reverse geocoding service}. Note that this service can only convert a single
set of coordinates at a time. See ArcGIS's
\href{https://developers.arcgis.com/rest/geocode/api-reference/
geocoding-reverse-geocode.htm}{Reverse Geocode} documentation page for
additional details.
}
\examples{
reverse_geocode(2516412.847819, 6986517.2125469996)

addresses <- data.frame(
  street = c('8525 Garland Rd', '1500 Marilla St', '3809 Grand Avenue'),
  city = c('Dallas', 'Dallas', 'Dallas'),
  zip = c('75218', '75201', '75210')
)

coords <- geocode_addresses(addresses$street, addresses$city, addresses$zip)
reverse_geocode(coords$latitude[[2]], coords$longitude[[2]]
}
